SIC/requirements.txt
torch
torchvision
numpy
pandas
psutil
jenkspy
matplotlib
pyyaml
dill
